chore: Update package.json with start and dev scripts
This commit adds the "start" and "dev" scripts to the package.json file. The "start" script is used to run the application using the "node index.js" command, while the "dev" script uses nodemon to automatically restart the server whenever changes are made to the code. This update improves the development workflow by providing a convenient way to start and develop the application.