refactor: Remove unnecessary files and configurations
This code change removes unnecessary files and configurations related to the `debug` module. The deleted files include `.coveralls.yml`, `.npmignore`, `.travis.yml`, `.eslintrc`, `component.json`, and some files in the `debug` and `ms` packages. This refactor improves the cleanliness and simplicity of the codebase by removing unused code and dependencies.